[changed] BitmapFont to use cap height as origin rather than baseline. Named methods better. Added method to get text bounds. [added] BitmapFontAlignmentTest.

+++ gdx/src/com/badlogic/gdx/graphics/particles/ParticleEffect.java
@@ -60,7 +60,7 @@ public class ParticleEffect {
 		}
 	}
 
-	public void setPosition (int x, int y) {
+	public void setPosition (float x, float y) {
 		for (int i = 0, n = emitters.size(); i < n; i++)
 			emitters.get(i).setPosition(x, y);
 	}
